Update TestUtilities.java
The error string used in validateCurrentRecord is assembled using the expected value twice, so you end up with something like "Value:'foo' did not match the expected value 'foo'" if the actual value 'bar'.  The fix updates that string to "Value:'bar' did not match the expected value 'foo'", which is a much more useful message.